fixup! ASoC: SOF: Implement snd_sof_bytes_ext_volatile_get kcontrol IO
fix compilation warning in i386 mode.

sound/soc/sof/control.c: In function ‘snd_sof_bytes_ext_volatile_get’:

sound/soc/sof/control.c:388:35: warning: format ‘%lu’ expects argument
of type ‘long unsigned int’, but argument 4 has type ‘unsigned int’
[-Wformat=]
  388 |   dev_err_ratelimited(scomp->dev, "error: user data size %d
  exceeds max size %lu.\n",
                     |
                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ sound/soc/sof/control.c
@@ -388,8 +388,9 @@ int snd_sof_bytes_ext_volatile_get(struct snd_kcontrol *kcontrol, unsigned int _
 
 	/* check data size doesn't exceed max coming from topology */
 	if (cdata->data->size > be->max - sizeof(const struct sof_abi_hdr)) {
-		dev_err_ratelimited(scomp->dev, "error: user data size %d exceeds max size %lu.\n",
-				    cdata->data->size, be->max - sizeof(const struct sof_abi_hdr));
+		dev_err_ratelimited(scomp->dev, "error: user data size %d exceeds max size %zu.\n",
+				    cdata->data->size,
+				    (size_t)be->max - sizeof(const struct sof_abi_hdr));
 		ret = -EINVAL;
 		goto out;
 	}
